Crate google_admin1_reports

This documentation was generated from reports crate version 0.1.8+20150115, where 20150115 is the exact revision of the admin:reports_v1 schema built by the mako code generator v0.1.8.

Everything else about the reports v1_reports API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

activities

list and watch

channels

stop

customer usage reports

get

user usage report

get

Subscription supported by ...

watch activities

list activities

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.activities().watch(...).doit()
let r = hub.channels().stop(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-admin1_reports = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_admin1_reports as admin1_reports;
use admin1_reports::Channel;
use admin1_reports::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use admin1_reports::Reports;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Reports::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Channel::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.activities().watch(req, "userKey", "applicationName")
             .start_time("sanctus")
             .page_token("et")
             .max_results(-46)
             .filters("et")
             .event_name("consetetur")
             .end_time("ut")
             .customer_id("ea")
             .actor_ip_address("sed")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Activities
JSON template for a collection of activites.

Activity
JSON template for the activity resource.

ActivityActor
User doing the action.

ActivityEvents
Activity events.

ActivityEventsParameters
Parameter value pairs for various applications.

ActivityId
Unique identifier for each activity record.

ActivityListCall
Retrieves a list of activities for a specific customer and application.

ActivityMethods
A builder providing access to all methods supported on activity resources. It is not used directly, but through the Reports hub.

ActivityWatchCall
Push changes to activities

Channel
An notification channel used to watch for resource changes.

ChannelMethods
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Reports hub.

ChannelStopCall
Stop watching resources through this channel

CustomerUsageReportGetCall
Retrieves a report which is a collection of properties / statistics for a specific customer.

CustomerUsageReportMethods
A builder providing access to all methods supported on customerUsageReport resources. It is not used directly, but through the Reports hub.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Reports
Central instance to access all Reports related resource activities

UsageReport
JSON template for a usage report.

UsageReportEntity
Information about the type of the item.

UsageReportParameters
Parameter value pairs for various applications.

UsageReports
JSON template for a collection of usage reports.

UsageReportsWarnings
Warnings if any.

UsageReportsWarningsData
Key-Value pairs to give detailed information on the warning.

UserUsageReportGetCall
Retrieves a report which is a collection of properties / statistics for a set of users.

UserUsageReportMethods
A builder providing access to all methods supported on userUsageReport resources. It is not used directly, but through the Reports hub.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.